+++ NEWS.md
@@ -9,6 +9,10 @@
 * The `n`-argument in `row_means()` and `row_sums()` now also may be `Inf`, to compute means or sums only if all values in a row are valid (i.e. non-missing).
 * Argument `weight.by` in `frq()` was renamed into `weights`.
 
+## Bug fixex
+
+* `round_num()` preserves data frame attributes.
+
 # sjmisc 2.7.4
 
 ## New functions